+++ CHANGELOG.md
@@ -16,6 +16,8 @@ on the default branch using [git-cliff](https://git-cliff.org/).
 
 ### Features
 
+- **metrics:** Bound profile label cardinality (EC-P2-09) [fc8a06c](https://github.com/konih/kollect/commit/fc8a06cabeaa205bef5c0672dfc721e56c795ad2)
+
 - **scope:** Cluster static-ref namespace allowlist (#23) ([#23](https://github.com/konih/kollect/pull/23))[c55f538](https://github.com/konih/kollect/commit/c55f5380d99ab9cf30dbd3f4122c18f6ff4fb3dc)